refactor: extract CreatedHistoryItem to a shared AgentInstanceCreatedHistoryItem client response type
Promotes CreatedHistoryItem from a type nested inside
UpdateAgentInstanceResponse to a top-level response type, renamed to
AgentInstanceCreatedHistoryItem, so it can be shared with the create
command's response without nesting it under an unrelated interface.
Pure move: no behavior change.
